@@ -1,9 +1,10 @@
 {
   "name": "react-diff-viewer",
-  "version": "1.1.0",
+  "version": "2.0.0",
   "private": false,
-  "description": "The only text diff viewer needed for React",
+  "description": "A simple and beautiful text diff viewer component made with diff and React",
   "keywords": [
+    "review",
     "code-review",
     "diff",
     "diff-viewer",
@@ -24,47 +25,55 @@
     "publish:examples": "NODE_ENV=production yarn build:examples && gh-pages -d examples/dist -r $GITHUB_REPO_URL",
     "publish:examples:local": "NODE_ENV=production yarn build:examples && gh-pages -d examples/dist",
     "start:examples": "webpack-dev-server --open --hot --inline",
-    "test": "mocha --require ts-node/register --require enzyme.ts ./test/index.tsx",
-    "test:watch": "mocha --require ts-node/register --require enzyme.ts --watch-extensions ts,tsx --watch ./test/index.tsx"
+    "test": "mocha --require ts-node/register --require enzyme.ts ./test/**",
+    "test:watch": "mocha --require ts-node/register --require enzyme.ts --watch-extensions ts,tsx --watch ./test/**"
   },
   "dependencies": {
     "classnames": "^2.2.6",
-    "diff": "^3.5.0",
-    "emotion": "^9.2.10",
-    "memoize-one": "^4.0.2",
+    "create-emotion": "^10.0.14",
+    "diff": "^4.0.1",
+    "emotion": "^10.0.14",
+    "memoize-one": "^5.0.4",
     "prop-types": "^15.6.2"
   },
   "devDependencies": {
     "@types/classnames": "^2.2.6",
-    "@types/diff": "^3.5.1",
+    "@types/diff": "^4.0.2",
     "@types/enzyme": "^3.1.14",
     "@types/enzyme-adapter-react-16": "^1.0.3",
     "@types/expect": "^1.20.3",
-    "@types/memoize-one": "^3.1.1",
+    "@types/memoize-one": "^4.1.1",
     "@types/mocha": "^5.2.5",
-    "@types/node": "^10.12.2",
+    "@types/node": "^12.0.12",
     "@types/react": "^16.4.14",
     "@types/react-dom": "^16.0.8",
     "@types/webpack": "^4.4.13",
-    "css-loader": "^1.0.0",
+    "@typescript-eslint/eslint-plugin": "^1.11.0",
+    "@typescript-eslint/parser": "^1.11.0",
+    "css-loader": "^3.0.0",
     "enzyme": "^3.7.0",
     "enzyme-adapter-react-16": "^1.6.0",
-    "expect": "^23.6.0",
+    "eslint": "6.0.1",
+    "eslint-config-airbnb": "17.1.1",
+    "eslint-plugin-import": "^2.18.0",
+    "eslint-plugin-jsx-a11y": "^6.2.3",
+    "eslint-plugin-react": "^7.14.2",
+    "expect": "^24.8.0",
     "favicons-webpack-plugin": "^0.0.9",
-    "file-loader": "^2.0.0",
+    "file-loader": "^4.0.0",
     "gh-pages": "^2.0.1",
     "html-webpack-plugin": "^3.2.0",
-    "mini-css-extract-plugin": "^0.4.3",
-    "mocha": "^5.2.0",
+    "mini-css-extract-plugin": "^0.7.0",
+    "mocha": "^6.1.4",
     "node-sass": "^4.9.3",
-    "raw-loader": "^0.5.1",
+    "raw-loader": "^3.0.0",
     "react": "^16.5.2",
     "react-dom": "^16.5.2",
     "sass-loader": "^7.1.0",
     "spy": "^1.0.0",
-    "ts-loader": "^5.2.1",
-    "ts-node": "^7.0.1",
-    "typescript": "^3.1.1",
+    "ts-loader": "^6.0.4",
+    "ts-node": "^8.3.0",
+    "typescript": "^3.5.2",
     "webpack": "^4.20.2",
     "webpack-cli": "^3.1.1",
     "webpack-dev-server": "^3.1.9"
