@@ -1,73 +1,68 @@
-import * as React from 'react'
-import * as diff from 'diff'
-import * as PropTypes from 'prop-types'
-import cn from 'classnames'
+import * as React from 'react';
+import * as PropTypes from 'prop-types';
+import cn from 'classnames';
 
-const m = require('memoize-one')
-const memoize = m.default || m
+import {
+  computeLineInformation,
+  LineInformation,
+  DiffInformation,
+  DiffType,
+} from './compute-lines';
+import computeStyles, { ReactDiffViewerStylesOverride, ReactDiffViewerStyles } from './styles';
 
-import computeStyles, {
-  IReactDiffViewerStyles,
-  IReactDiffViewerStylesOverride,
-} from './styles'
-import { InlineLine, DefaultLine } from './line'
+// eslint-disable-next-line @typescript-eslint/no-var-requires
+const m = require('memoize-one');
 
-export interface IReactDiffViewerProps {
+const memoize = m.default || m;
+
+export enum LineNumberPrefix {
+  LEFT = 'L',
+  RIGHT = 'R',
+}
+
+export interface ReactDiffViewerProps {
+  // Old value to compare.
   oldValue: string;
+  // New value to compare.
   newValue: string;
+  // Enable/Disable split view.
   splitView?: boolean;
+  // Enable/Disable word diff.
   disableWordDiff?: boolean;
+  // Number of unmodified lines surrounding each line diff.
+  extraLinesSurroundingDiff?: number;
+  // Show/hide line number.
   hideLineNumbers?: boolean;
+  // Show only diff between the two values.
+  showDiffOnly?: boolean;
+  // Render prop to format final string before displaying them in the UI.
   renderContent?: (source: string) => JSX.Element;
-  onLineNumberClick?: (lineId: string, event: React.MouseEvent<HTMLTableCellElement>) => void;
+  // Render prop to format code fold message.
+  codeFoldMessageRenderer?: (
+    totalFoldedLines: number,
+    leftStartLineNumber: number,
+    rightStartLineNumber: number,
+  ) => JSX.Element;
+  // Event handler for line number click.
+  onLineNumberClick?: (
+    lineId: string,
+    event: React.MouseEvent<HTMLTableCellElement>,
+  ) => void;
+  // Array of line ids to highlight lines.
   highlightLines?: string[];
-  styles?: IReactDiffViewerStylesOverride,
+  // Style overrides.
+  styles?: ReactDiffViewerStylesOverride;
 }
 
-export interface IReactDiffViewerState {
-
+export interface ReactDiffViewerState {
+  // Array holding the expanded code folding.
+  expandedBlocks?: number[];
 }
 
-const wordDiff = (
-  oldValue: string,
-  newValue: string,
-  hideType: string,
-  styles: IReactDiffViewerStyles,
-  renderContent?: (source: string) => JSX.Element,
-) => {
-  const charDiff = diff.diffWordsWithSpace(oldValue, newValue)
-  return charDiff.map((obj: any, i) => {
-    if (obj[hideType]) return undefined
-    if (renderContent) {
-      return <span
-        className={cn(
-          styles.wordDiff,
-          {
-            [styles.wordAdded]: obj.added,
-            [styles.wordRemoved]: obj.removed,
-          },
-        )}
-        key={i}>
-        { renderContent(obj.value) }
-      </span>
-    }
-    return <pre
-      className={cn(
-        styles.wordDiff,
-        {
-          [styles.wordAdded]: obj.added,
-          [styles.wordRemoved]: obj.removed,
-        },
-      )}
-      key={i}>
-      { obj.value }
-    </pre>
-  })
-}
+class DiffViewer extends React.Component<ReactDiffViewerProps, ReactDiffViewerState> {
+  private styles: ReactDiffViewerStyles;
 
-class DiffViewer extends React.Component<IReactDiffViewerProps, IReactDiffViewerState> {
-
-  static defaultProps: IReactDiffViewerProps = {
+  public static defaultProps: ReactDiffViewerProps = {
     oldValue: '',
     newValue: '',
     splitView: true,
@@ -75,185 +70,433 @@ class DiffViewer extends React.Component<IReactDiffViewerProps, IReactDiffViewer
     disableWordDiff: false,
     styles: {},
     hideLineNumbers: false,
-  }
+    extraLinesSurroundingDiff: 3,
+    showDiffOnly: true,
+  };
 
-  static propTypes = {
+  public static propTypes = {
     oldValue: PropTypes.string.isRequired,
     newValue: PropTypes.string.isRequired,
     splitView: PropTypes.bool,
     disableWordDiff: PropTypes.bool,
     renderContent: PropTypes.func,
     onLineNumberClick: PropTypes.func,
+    extraLinesSurroundingDiff: PropTypes.number,
     styles: PropTypes.object,
     hideLineNumbers: PropTypes.bool,
+    showDiffOnly: PropTypes.bool,
     highlightLines: PropTypes.arrayOf(PropTypes.string),
+  };
+
+  public constructor(props: ReactDiffViewerProps) {
+    super(props);
+
+    this.state = {
+      expandedBlocks: [],
+    };
   }
 
-  private splitView = (diffArray: diff.IDiffResult[], styles: IReactDiffViewerStyles) => {
-    let leftLineNumber = 0
-    let rightLineNumber = 0
-
-    return () => diffArray.map((obj: diff.IDiffResult, i) => {
-      return <React.Fragment key={i}>
-        {
-          obj.value.split('\n')
-            .filter(ch => ch.length > 0)
-            .map((ch, num) => {
-              if (!obj.added && !obj.removed) {
-                rightLineNumber = rightLineNumber + 1
-                leftLineNumber = leftLineNumber + 1
-                return <DefaultLine
-                  styles={styles}
-                  hideLineNumbers={this.props.hideLineNumbers}
-                  leftLineNumber={leftLineNumber}
-                  rightLineNumber={rightLineNumber}
-                  leftContent={ch}
-                  rightContent={ch}
-                  key={num}
-                  hightlightLines={this.props.highlightLines}
-                  renderContent={this.props.renderContent}
-                  onLineNumberClick={this.props.onLineNumberClick}
-                />
-              }
-
-              let leftContent
-              let rightContent
-              let removed = obj.removed
-              let added = obj.added
-              if (obj.added && diffArray[i - 1] && diffArray[i - 1].removed) {
-                const preValueCount = diffArray[i - 1].count
-                if (num <= (preValueCount - 1)) return undefined
-                rightLineNumber = rightLineNumber + 1
-                rightContent = ch
-              } else if (obj.removed && diffArray[i + 1] && !diffArray[i + 1].added) {
-                leftLineNumber = leftLineNumber + 1
-                leftContent = ch
-              } else if (obj.removed && diffArray[i + 1] && diffArray[i + 1].added) {
-                leftLineNumber = leftLineNumber + 1
-                const nextVal = diffArray[i + 1].value
-                  .split('\n')
-                  .filter(Boolean)[num]
-                leftContent = (nextVal && !this.props.disableWordDiff)
-                  ? wordDiff(ch, nextVal, 'added', styles, this.props.renderContent)
-                  : ch
-                rightContent = (nextVal && !this.props.disableWordDiff)
-                  ? wordDiff(ch, nextVal, 'removed', styles, this.props.renderContent)
-                  : nextVal
-                if (nextVal) {
-                  rightLineNumber = rightLineNumber + 1
-                  added = true
-                }
-              } else {
-                rightLineNumber = rightLineNumber + 1
-                rightContent = ch
-              }
-              return <DefaultLine
-                styles={styles}
-                leftLineNumber={!removed || leftLineNumber}
-                rightLineNumber={!added || rightLineNumber}
-                removed={removed}
-                added={added}
-                key={num}
-                hideLineNumbers={this.props.hideLineNumbers}
-                hightlightLines={this.props.highlightLines}
-                renderContent={this.props.renderContent}
-                leftContent={leftContent}
-                rightContent={rightContent}
-                onLineNumberClick={this.props.onLineNumberClick}
-              />
-            })
-        }
-      </React.Fragment>
-    })
-  }
-
-  private inlineView = (diffArray: diff.IDiffResult[], styles: IReactDiffViewerStyles) => {
-    let leftLineNumber = 0
-    let rightLineNumber = 0
-    return () => {
-      return diffArray.map((diffObj, i) => {
-        return diffObj.value.split('\n')
-          .filter(ch => ch.length > 0)
-          .map((ch, num) => {
-            let content
-            if (diffObj.added) {
-              rightLineNumber = rightLineNumber + 1
-              if (diffArray[i - 1] && diffArray[i - 1].removed) {
-                const preValue = diffArray[i - 1].value
-                .split('\n')
-                .filter(Boolean)[num]
-                content = (preValue && !this.props.disableWordDiff) ? wordDiff(preValue, ch, 'removed', styles, this.props.renderContent) : ch
-              } else {
-                content = ch
-              }
-            } else if (diffObj.removed) {
-              leftLineNumber = leftLineNumber + 1
-              if (diffArray[i + 1] && diffArray[i + 1].added) {
-                const nextVal = diffArray[i + 1].value
-                .split('\n')
-                .filter(Boolean)[num]
-                content = (nextVal && !this.props.disableWordDiff) ? wordDiff(ch, nextVal, 'added', styles, this.props.renderContent) : ch
-              } else {
-                content = ch
-              }
-            } else {
-              rightLineNumber = rightLineNumber + 1
-              leftLineNumber = leftLineNumber + 1
-              content = ch
-            }
-            return <InlineLine
-              styles={styles}
-              onLineNumberClick={this.props.onLineNumberClick}
-              key={num}
-              hideLineNumbers={this.props.hideLineNumbers}
-              renderContent={this.props.renderContent}
-              removed={diffObj.removed}
-              leftLineNumber={diffObj.added || leftLineNumber}
-              rightLineNumber={diffObj.removed || rightLineNumber}
-              content={content}
-              hightlightLines={this.props.highlightLines}
-              added={diffObj.added}
-            />
-          })
-      })
+  /**
+   * Resets code block expand to the initial stage. Will be exposed to the parent component via
+   * refs.
+   */
+  public resetCodeBlocks = (): boolean => {
+    if (this.state.expandedBlocks.length > 0) {
+      this.setState({
+        expandedBlocks: [],
+      });
+      return true;
     }
+    return false;
   }
 
-  private computeStyles = memoize(computeStyles)
+  /**
+   * Pushes the target expanded code block to the state. During the re-render,
+   * this value is used to expand/fold unmodified code.
+   */
+  private onBlockExpand = (id: number): void => {
+    const prevState = this.state.expandedBlocks.slice();
+    prevState.push(id);
 
-  public render = () => {
+    this.setState({
+      expandedBlocks: prevState,
+    });
+  };
 
-    const {
-      oldValue,
-      newValue,
-      splitView,
-    } = this.props
+  /**
+   * Computes final styles for the diff viewer. It combines the default styles with the user
+   * supplied overrides. The computed styles are cached with performance in mind.
+   *
+   * @param styles User supplied style overrides.
+   */
+  private computeStyles: (
+    styles: ReactDiffViewerStylesOverride,
+  ) => ReactDiffViewerStyles = memoize(computeStyles);
 
-    if (typeof oldValue !== 'string' || typeof newValue !== 'string') {
-      throw Error('"oldValue" and "newValue" should be strings')
+  /**
+   * Returns a function with clicked line number in the closure. Returns an no-op function when no
+   * onLineNumberClick handler is supplied.
+   *
+   * @param id Line id of a line.
+   */
+  private onLineNumberClickProxy = (id: string): any => {
+    if (this.props.onLineNumberClick) {
+      return (e: any): void => this.props.onLineNumberClick(id, e);
     }
+    return (): void => { };
+  };
 
-    const newStyles: any = this.computeStyles(this.props.styles)
+  /**
+   * Maps over the word diff and constructs the required React elements to show word diff.
+   *
+   * @param diffArray Word diff information derived from line information.
+   * @param renderer Optional renderer to format diff words. Useful for syntax highlighting.
+   */
+  private renderWordDiff = (
+    diffArray: DiffInformation[],
+    renderer?: (chunk: string) => JSX.Element,
+  ): JSX.Element[] => {
+    return diffArray.map(
+      (wordDiff, i): JSX.Element => {
+        return (
+          <span
+            key={i}
+            className={cn(this.styles.wordDiff, {
+              [this.styles.wordAdded]: wordDiff.type === DiffType.ADDED,
+              [this.styles.wordRemoved]: wordDiff.type === DiffType.REMOVED,
+            })}
+          >
+            {renderer ? renderer(wordDiff.value as string) : wordDiff.value}
+          </span>
+        );
+      },
+    );
+  };
 
-    const diffArray = diff.diffLines(oldValue, newValue, {
-      newlineIsToken: true,
-      ignoreWhitespace: false,
-      ignoreCase: false,
-    })
-
-    const nodes = splitView
-      ? this.splitView(diffArray, newStyles)()
-      : this.inlineView(diffArray, newStyles)()
+  /**
+   * Maps over the line diff and constructs the required react elements to show line diff. It calls
+   * renderWordDiff when encountering word diff. This takes care of both inline and split view line
+   * renders.
+   *
+   * @param lineNumber Line number of the current line.
+   * @param type Type of diff of the current line.
+   * @param prefix Unique id to prefix with the line numbers.
+   * @param value Content of the line. It can be a string or a word diff array.
+   * @param additionalLineNumber Additional line number to be shown. Useful for rendering inline
+   *  diff view. Right line number will be passed as additionalLineNumber.
+   * @param additionalPrefix Similar to prefix but for additional line number.
+   */
+  private renderLine = (
+    lineNumber: number,
+    type: DiffType,
+    prefix: LineNumberPrefix,
+    value: string | DiffInformation[],
+    additionalLineNumber?: number,
+    additionalPrefix?: LineNumberPrefix,
+  ): JSX.Element => {
+    const lineNumberTemplate = `${prefix}-${lineNumber}`;
+    const additionalLineNumberTemplate = `${additionalPrefix}-${additionalLineNumber}`;
+    const highlightLine = this.props.highlightLines.includes(lineNumberTemplate)
+      || this.props.highlightLines.includes(additionalLineNumberTemplate);
+    const added = type === DiffType.ADDED;
+    const removed = type === DiffType.REMOVED;
+    let content;
+    if (Array.isArray(value)) {
+      content = this.renderWordDiff(value, this.props.renderContent);
+    } else if (this.props.renderContent) {
+      content = this.props.renderContent(value);
+    } else {
+      content = value;
+    }
 
     return (
-      <table className={newStyles.diffContainer}>
-        <tbody>
-          {nodes}
-        </tbody>
-      </table>
-    )
+      <React.Fragment>
+        {!this.props.hideLineNumbers && (
+          <td
+            onClick={
+              lineNumber && this.onLineNumberClickProxy(lineNumberTemplate)
+            }
+            className={cn(this.styles.gutter, {
+              [this.styles.diffAdded]: added,
+              [this.styles.diffRemoved]: removed,
+              [this.styles.highlightedGutter]: highlightLine,
+            })}
+          >
+            <pre>{lineNumber}</pre>
+          </td>
+        )}
+        {!this.props.splitView && !this.props.hideLineNumbers && (
+          <td
+            onClick={
+              additionalLineNumber
+              && this.onLineNumberClickProxy(additionalLineNumberTemplate)
+            }
+            className={cn(this.styles.gutter, {
+              [this.styles.diffAdded]: added,
+              [this.styles.diffRemoved]: removed,
+              [this.styles.highlightedGutter]: highlightLine,
+            })}
+          >
+            <pre>{additionalLineNumber}</pre>
+          </td>
+        )}
+        <td
+          className={cn(this.styles.marker, {
+            [this.styles.emptyLine]: !content,
+            [this.styles.diffAdded]: added,
+            [this.styles.diffRemoved]: removed,
+            [this.styles.highlightedLine]: highlightLine,
+          })}
+        >
+          <pre>
+            {added && '+'}
+            {removed && '-'}
+          </pre>
+        </td>
+        <td
+          className={cn({
+            [this.styles.emptyLine]: !content,
+            [this.styles.diffAdded]: added,
+            [this.styles.diffRemoved]: removed,
+            [this.styles.highlightedLine]: highlightLine,
+          })}
+        >
+          <pre>{content}</pre>
+        </td>
+      </React.Fragment>
+    );
+  };
+
+  /**
+   * Generates lines for split view.
+   *
+   * @param obj Line diff information.
+   * @param obj.left Life diff information for the left pane of the split view.
+   * @param obj.right Life diff information for the right pane of the split view.
+   * @param index React key for the lines.
+   */
+  private renderSplitView = (
+    { left, right }: LineInformation,
+    index: number,
+  ): JSX.Element => {
+    return (
+      <tr key={index} className={this.styles.line}>
+        {this.renderLine(
+          left.lineNumber,
+          left.type,
+          LineNumberPrefix.LEFT,
+          left.value,
+        )}
+        {this.renderLine(
+          right.lineNumber,
+          right.type,
+          LineNumberPrefix.RIGHT,
+          right.value,
+        )}
+      </tr>
+    );
+  };
+
+  /**
+   * Generates lines for inline view.
+   *
+   * @param obj Line diff information.
+   * @param obj.left Life diff information for the added section of the inline view.
+   * @param obj.right Life diff information for the removed section of the inline view.
+   * @param index React key for the lines.
+   */
+  public renderInlineView = (
+    { left, right }: LineInformation,
+    index: number,
+  ): JSX.Element => {
+    let content;
+    if (left.type === DiffType.REMOVED && right.type === DiffType.ADDED) {
+      return (
+        <React.Fragment key={index}>
+          <tr className={this.styles.line}>
+            {this.renderLine(
+              left.lineNumber,
+              left.type,
+              LineNumberPrefix.LEFT,
+              left.value,
+              null,
+            )}
+          </tr>
+          <tr className={this.styles.line}>
+            {this.renderLine(
+              null,
+              right.type,
+              LineNumberPrefix.RIGHT,
+              right.value,
+              right.lineNumber,
+            )}
+          </tr>
+        </React.Fragment>
+      );
+    }
+    if (left.type === DiffType.REMOVED) {
+      content = this.renderLine(
+        left.lineNumber,
+        left.type,
+        LineNumberPrefix.LEFT,
+        left.value,
+        null,
+      );
+    }
+    if (left.type === DiffType.DEFAULT) {
+      content = this.renderLine(
+        left.lineNumber,
+        left.type,
+        LineNumberPrefix.LEFT,
+        left.value,
+        right.lineNumber,
+        LineNumberPrefix.RIGHT,
+      );
+    }
+    if (right.type === DiffType.ADDED) {
+      content = this.renderLine(
+        null,
+        right.type,
+        LineNumberPrefix.RIGHT,
+        right.value,
+        right.lineNumber,
+      );
+    }
+
+    return <tr key={index} className={this.styles.line}>{content}</tr>;
+  };
+
+  /**
+   * Returns a function with clicked block number in the closure.
+   *
+   * @param id Cold fold block id.
+   */
+  private onBlockClickProxy = (id: number): any => (): void => this.onBlockExpand(id);
+
+  /**
+   * Generates cold fold block. It also uses the custom message renderer when available to show
+   * cold fold messages.
+   *
+   * @param num Number of skipped lines between two blocks.
+   * @param blockNumber Code fold block id.
+   * @param leftBlockLineNumber First left line number after the current code fold block.
+   * @param rightBlockLineNumber First right line number after the current code fold block.
+   */
+  private renderSkippedLineIndicator = (
+    num: number,
+    blockNumber: number,
+    leftBlockLineNumber: number,
+    rightBlockLineNumber: number,
+  ): JSX.Element => {
+    const { splitView } = this.props;
+    const message = this.props.codeFoldMessageRenderer
+      ? this.props
+        .codeFoldMessageRenderer(num, leftBlockLineNumber, rightBlockLineNumber)
+      : <pre>Expand {num} lines ...</pre>;
+    const content = (
+      <td>
+        <a onClick={this.onBlockClickProxy(blockNumber)} tabIndex={0}>
+          {message}
+        </a>
+      </td>
+    );
+    return (
+      <tr key={Math.round(100).toString()} className={this.styles.codeFold}>
+        {!this.props.hideLineNumbers && (
+          <td className={this.styles.codeFoldGutter} />
+        )}
+        <td className={cn({ [this.styles.codeFoldGutter]: !splitView })} />
+        {splitView ? content : <td />}
+        {!splitView ? content : <td />}
+        <td />
+        <td />
+      </tr>
+    );
+  };
+
+  /**
+   * Generates the entire diff view.
+   */
+  private renderDiff = (): JSX.Element[] => {
+    const { oldValue, newValue, splitView } = this.props;
+    const { lineInformation, diffLines } = computeLineInformation(
+      oldValue,
+      newValue,
+      this.props.disableWordDiff,
+    );
+    const extraLines = this.props.extraLinesSurroundingDiff < 0
+      ? 0
+      : this.props.extraLinesSurroundingDiff;
+    let skippedLines: number[] = [];
+    return lineInformation.map(
+      (line: LineInformation, i: number): JSX.Element => {
+        const diffBlockStart = diffLines[0];
+        const currentPosition = diffBlockStart - i;
+        if (this.props.showDiffOnly) {
+          if (currentPosition === -extraLines) {
+            skippedLines = [];
+            diffLines.shift();
+          }
+          if (
+            line.left.type === DiffType.DEFAULT
+            && (currentPosition > extraLines
+              || typeof diffBlockStart === 'undefined')
+            && !this.state.expandedBlocks.includes(diffBlockStart)
+          ) {
+            skippedLines.push(i + 1);
+            if (i === lineInformation.length - 1 && skippedLines.length > 1) {
+              return this.renderSkippedLineIndicator(
+                skippedLines.length,
+                diffBlockStart,
+                line.left.lineNumber,
+                line.right.lineNumber,
+              );
+            }
+            return null;
+          }
+        }
+
+        const diffNodes = splitView
+          ? this.renderSplitView(line, i)
+          : this.renderInlineView(line, i);
+
+        if (currentPosition === extraLines && skippedLines.length > 1) {
+          const { length } = skippedLines;
+          skippedLines = [];
+          return (
+            <React.Fragment key={i}>
+              {this.renderSkippedLineIndicator(
+                length,
+                diffBlockStart,
+                line.left.lineNumber,
+                line.right.lineNumber,
+              )}
+              {diffNodes}
+            </React.Fragment>
+          );
+        }
+        return diffNodes;
+      },
+    );
   }
+
+  public render = (): JSX.Element => {
+    const { oldValue, newValue } = this.props;
+
+    if (typeof oldValue !== 'string' || typeof newValue !== 'string') {
+      throw Error('"oldValue" and "newValue" should be strings');
+    }
+
+    this.styles = this.computeStyles(this.props.styles);
+    const nodes = this.renderDiff();
+
+    return (
+      <table className={this.styles.diffContainer}>
+        <tbody>{nodes}</tbody>
+      </table>
+    );
+  };
 }
 
-export default DiffViewer
-export { IReactDiffViewerStylesOverride }
+export default DiffViewer;
+export { ReactDiffViewerStylesOverride };
