@@ -1,61 +1,67 @@
-import { css, cx } from 'emotion'
-import { Interpolation } from 'create-emotion'
+import { css, cx } from 'emotion';
+import { Interpolation } from 'create-emotion';
 
-
-export interface IReactDiffViewerStyles {
+export interface ReactDiffViewerStyles {
   diffContainer?: string;
   diffRemoved?: string;
   diffAdded?: string;
   line?: string;
-  hightlightedGutter?: string;
+  highlightedGutter?: string;
   gutter?: string;
-  hightlightedLine?: string;
+  highlightedLine?: string;
   marker?: string;
   wordDiff?: string;
   wordAdded?: string;
   wordRemoved?: string;
-  leftGutter?: string;
-  rightGutter?: string;
+  codeFoldGutter?: string;
+  emptyLine?: string;
+  codeFold?: string;
+  [key: string]: string;
 }
 
-export interface IReactDiffViewerStylesOverride {
+export interface ReactDiffViewerStylesOverride {
   variables?: {
-    addedBackground?: string,
-    addedColor?: string,
-    removedBackground?: string,
-    removedColor?: string,
-    wordAddedBackground?: string,
-    wordRemovedBackground?: string,
-    addedGutterBackground?: string,
-    removedGutterBackground?: string,
-    gutterBackground?: string,
-    gutterBackgroundDark?: string,
-    highlightBackground?: string,
-    highlightGutterBackground?: string,
-  },
+    diffViewerBackground?: string;
+    addedBackground?: string;
+    addedColor?: string;
+    removedBackground?: string;
+    removedColor?: string;
+    wordAddedBackground?: string;
+    wordRemovedBackground?: string;
+    addedGutterBackground?: string;
+    removedGutterBackground?: string;
+    gutterBackground?: string;
+    gutterBackgroundDark?: string;
+    highlightBackground?: string;
+    highlightGutterBackground?: string;
+    codeFoldGutterBackground?: string;
+    codeFoldBackground?: string;
+    emptyLineBackground?: string;
+  };
   diffContainer?: Interpolation;
   diffRemoved?: Interpolation;
   diffAdded?: Interpolation;
   marker?: Interpolation;
-  hightlightedLine?: Interpolation;
-  hightlightedGutter?: Interpolation;
+  highlightedLine?: Interpolation;
+  highlightedGutter?: Interpolation;
   gutter?: Interpolation;
   line?: Interpolation;
   wordDiff?: Interpolation;
   wordAdded?: Interpolation;
   wordRemoved?: Interpolation;
-  leftGutter?: Interpolation;
-  rightGutter?: Interpolation;
+  codeFoldGutter?: Interpolation;
+  emptyLine?: Interpolation;
 }
 
-export default (styleOverride: IReactDiffViewerStylesOverride) => {
+export default (styleOverride: ReactDiffViewerStylesOverride): ReactDiffViewerStyles => {
   const {
     variables: overrideVariables,
     ...styles
-  } = styleOverride
+  } = styleOverride;
 
   const variables = {
     ...{
+      diffViewerBackground: '#fff',
       addedBackground: '#e6ffed',
       addedColor: '#24292e',
       removedBackground: '#ffeef0',
@@ -68,33 +74,23 @@ export default (styleOverride: IReactDiffViewerStylesOverride) => {
       gutterBackgroundDark: '#f3f1f1',
       highlightBackground: '#fffbdd',
       highlightGutterBackground: '#fff5b1',
+      codeFoldGutterBackground: '#dbedff',
+      codeFoldBackground: '#f1f8ff',
+      emptyLineBackground: '#fafbfc',
     },
-    ...overrideVariables
-  }
+    ...overrideVariables,
+  };
 
   const diffContainer = css({
     width: '100%',
-    'pre': {
+    background: variables.diffViewerBackground,
+    pre: {
       margin: 0,
       whiteSpace: 'pre-wrap',
       lineHeight: '25px',
     },
-    'tbody': {
-      tr: {
-        '&:first-child': {
-          td: {
-            paddingTop: 15,
-          },
-        },
-        '&:last-child': {
-          td: {
-            paddingBottom: 15,
-          },
-        },
-      },
-    },
     label: 'diff-container',
-  })
+  });
 
   const diffRemoved = css({
     background: variables.removedBackground,
@@ -103,7 +99,7 @@ export default (styleOverride: IReactDiffViewerStylesOverride) => {
       color: variables.removedColor,
     },
     label: 'diff-removed',
-  })
+  });
 
   const diffAdded = css({
     background: variables.addedBackground,
@@ -112,33 +108,58 @@ export default (styleOverride: IReactDiffViewerStylesOverride) => {
       color: variables.addedColor,
     },
     label: 'diff-added',
-  })
+  });
+
+  const codeFoldGutter = css({
+    backgroundColor: variables.codeFoldGutterBackground,
+    label: 'code-fold-gutter',
+  });
+
+  const codeFold = css({
+    backgroundColor: variables.codeFoldBackground,
+    height: 40,
+    fontSize: 14,
+    fontWeight: 700,
+    label: 'code-fold',
+    a: {
+      textDecoration: 'underline !important',
+      cursor: 'pointer',
+      pre: {
+        display: 'inline',
+      },
+    },
+  });
+
+  const emptyLine = css({
+    backgroundColor: variables.emptyLineBackground,
+    label: 'empty-line',
+  });
 
   const marker = css({
-    minWidth: 50,
+    width: 25,
     paddingLeft: 10,
     userSelect: 'none',
     label: 'marker',
     [`&.${diffAdded}`]: {
       pre: {
         color: variables.addedColor,
-      }
+      },
     },
     [`&.${diffRemoved}`]: {
       pre: {
         color: variables.removedColor,
-      }
-    }
-  })
+      },
+    },
+  });
 
-  const hightlightedLine = css({
+  const highlightedLine = css({
     background: variables.highlightBackground,
-    label: 'hightlighted-line',
-  })
+    label: 'highlighted-line',
+  });
 
-  const hightlightedGutter = css({
-    label: 'hightlighted-gutter',
-  })
+  const highlightedGutter = css({
+    label: 'highlighted-gutter',
+  });
 
   const gutter = css({
     userSelect: 'none',
@@ -163,75 +184,68 @@ export default (styleOverride: IReactDiffViewerStylesOverride) => {
     [`&.${diffRemoved}`]: {
       background: variables.removedGutterBackground,
     },
-    [`&.${hightlightedGutter}`]: {
+    [`&.${highlightedGutter}`]: {
       background: variables.highlightGutterBackground,
       '&:hover': {
         background: variables.highlightGutterBackground,
       },
     },
-  })
+  });
 
   const line = css({
     verticalAlign: 'baseline',
     label: 'line',
-  })
+  });
 
   const wordDiff = css({
     padding: 2,
     display: 'inline-flex',
     borderRadius: 1,
     label: 'word-diff',
-  })
+  });
 
   const wordAdded = css({
     background: variables.wordAddedBackground,
     label: 'word-added',
-  })
+  });
 
   const wordRemoved = css({
     background: variables.wordRemovedBackground,
     label: 'word-removed',
-  })
-
-  const leftGutter = css({
-    label: 'left-gutter',
-  })
-
-  const rightGutter = css({
-    label: 'right-gutter',
-  })
+  });
 
   const defaultStyles: any = {
     diffContainer,
     diffRemoved,
     diffAdded,
     marker,
-    hightlightedGutter,
-    hightlightedLine,
+    highlightedGutter,
+    highlightedLine,
     gutter,
     line,
     wordDiff,
     wordAdded,
     wordRemoved,
-    leftGutter,
-    rightGutter,
-  }
+    codeFoldGutter,
+    codeFold,
+    emptyLine,
+  };
 
-  const computerOverrideStyles: any = Object.keys(styles)
-    .reduce((acc, key) => ({
+  const computerOverrideStyles: ReactDiffViewerStyles = Object.keys(styles)
+    .reduce((acc, key): ReactDiffViewerStyles => ({
       ...acc,
       ...{
         [key]: css((styles as any)[key]),
-      }
-    }), {})
+      },
+    }), {});
 
   return Object.keys(defaultStyles)
-    .reduce((acc, key) => ({
+    .reduce((acc, key): ReactDiffViewerStyles => ({
       ...acc,
       ...{
         [key]: computerOverrideStyles[key]
           ? cx(defaultStyles[key], computerOverrideStyles[key])
           : defaultStyles[key],
       },
-    }), {})
-}
+    }), {});
+};
